DobbyTourGuide/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
constantly==23.10.4
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
elevenlabs==1.0.5
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
frozenlist==1.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
numpy==1.26.4
openai==0.28.0
osrf-pycommon==2.0.2
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pip-upgrader==1.4.15
platformdirs==4.2.0
playsound==1.3.0
pocketsphinx==5.0.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==5.26.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pygame==2.5.2
Pygments==2.17.2
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
roslibpy==1.6.0
rpds-py==0.18.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.3
terminaltables==3.1.10
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformations==2024.6.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
zope.interface==6.3